+++ tests/tests_pytorch/trainer/connectors/test_data_connector.py
@@ -526,19 +526,20 @@ def test_invalid_hook_passed_in_datahook_selector():
         dh_selector.get_instance("setup")
 
 
-def test_eval_distributed_sampler_warning(tmpdir):
+@pytest.mark.parametrize("devices, warn_context", [(1, no_warning_call), (2, pytest.warns)])
+def test_eval_distributed_sampler_warning(devices, warn_context):
     """Test that a warning is raised when `DistributedSampler` is used with evaluation."""
 
     model = BoringModel()
-    trainer = Trainer(strategy="ddp", devices=2, accelerator="cpu", fast_dev_run=True)
+    trainer = Trainer(strategy="ddp", devices=devices, accelerator="cpu")
     trainer._data_connector.attach_data(model)
 
     trainer.state.fn = TrainerFn.VALIDATING
-    with pytest.warns(PossibleUserWarning, match="multi-device settings use `DistributedSampler`"):
+    with warn_context(PossibleUserWarning, match="multi-device settings use `DistributedSampler`"):
         trainer.reset_val_dataloader(model)
 
     trainer.state.fn = TrainerFn.TESTING
-    with pytest.warns(PossibleUserWarning, match="multi-device settings use `DistributedSampler`"):
+    with warn_context(PossibleUserWarning, match="multi-device settings use `DistributedSampler`"):
         trainer.reset_test_dataloader(model)